Move version check method location
