+++ sdk/core/azure-core-http-netty/src/main/java/com/azure/core/http/netty/NettyAsyncHttpClientBuilder.java
@@ -9,7 +9,7 @@ import com.azure.core.http.netty.implementation.AzureNettyHttpClientContext;
 import com.azure.core.http.netty.implementation.AzureSdkHandler;
 import com.azure.core.http.netty.implementation.ChallengeHolder;
 import com.azure.core.http.netty.implementation.HttpProxyHandler;
-import com.azure.core.http.netty.implementation.NettyVersion;
+import com.azure.core.http.netty.implementation.Utility;
 import com.azure.core.util.AuthorizationChallengeHandler;
 import com.azure.core.util.Configuration;
 import com.azure.core.util.Context;
@@ -86,7 +86,7 @@ public class NettyAsyncHttpClientBuilder {
         DEFAULT_READ_TIMEOUT = getDefaultTimeoutFromEnvironment(configuration, PROPERTY_AZURE_REQUEST_READ_TIMEOUT,
             Duration.ofSeconds(60), LOGGER).toMillis();
 
-        NettyVersion.validateNettyVersions();
+        Utility.validateNettyVersions();
     }
 
     private final HttpClient baseHttpClient;
